Finish moving mappers to src/lib
