+++ src/modules/billing/mappers/index.js
@@ -10,7 +10,6 @@ exports.contact = require('./contact');
 exports.invoice = require('./invoice');
 exports.invoiceAccount = require('./invoice-account');
 exports.invoiceLicence = require('./invoice-licence');
-exports.licence = require('./licence');
-exports.region = require('./region');
+exports.licence = require('../../../lib/mappers/licence');
 exports.transaction = require('./transaction');
 exports.user = require('./user');
