+++ test/modules/licences/controllers/documents.js
@@ -1,19 +1,19 @@
 const { expect } = require('@hapi/code');
 const { afterEach, beforeEach, experiment, test } = exports.lab = require('@hapi/lab').script();
 
-const controller = require('../../../src/modules/licences/controller');
-const queries = require('../../../src/modules/licences/lib/queries');
-const permitClient = require('../../../src/lib/connectors/permit');
-const documentsClient = require('../../../src/lib/connectors/crm/documents');
-const crmEntities = require('../../../src/lib/connectors/crm/entities');
-const idmConnector = require('../../../src/lib/connectors/idm');
-const { logger } = require('../../../src/logger');
+const controller = require('../../../../src/modules/licences/controllers/documents');
+const queries = require('../../../../src/modules/licences/lib/queries');
+const permitClient = require('../../../../src/lib/connectors/permit');
+const documentsClient = require('../../../../src/lib/connectors/crm/documents');
+const crmEntities = require('../../../../src/lib/connectors/crm/entities');
+const idmConnector = require('../../../../src/lib/connectors/idm');
+const { logger } = require('../../../../src/logger');
 
 const { cloneDeep } = require('lodash');
 const sinon = require('sinon');
 const sandbox = sinon.createSandbox();
 
-const { licences } = require('../../responses/permits/licence');
+const { licences } = require('../../../responses/permits/licence');
 
 const testRequest = {
   params: {
